+++ wenlv/src/main/java/com/zte/rest/NewsRest.java
@@ -173,25 +173,25 @@ public class NewsRest {
 			}else {
 				record.setUpdatetime(null);
 				record.setUpdater(admin.getAdminId());
-				//如果有标签id列表
-				if (labelIds.size()!=0){
-					Integer resourceid = record.getResourceid();
-					if (resourceid==null||resourceNewsService.selectByPrimaryKey(resourceid)==null){
-						return JsonResult.getFail("未获取到正确的resourceid");
-					}
-					ResourceLabelVo resourceLabelVo = new ResourceLabelVo();
-					resourceLabelVo.setCreator(admin.getAdminId());
-					resourceLabelVo.setUpdater(admin.getAdminId());
-					resourceLabelVo.setCreatetime(currentTime);
-					resourceLabelVo.setUpdatetime(currentTime);
-					resourceLabelVo.setTableid(2);
-					resourceLabelVo.setResourceid(resourceid);
-					resourceLabelVo.setStatus(1);
-					resourceLabelService.deleteResourceLabelByResourceId(resourceid);
-					for (Integer labelId : labelIds) {
-						resourceLabelVo.setLabelid(labelId);
-						resourceLabelService.addResourceLabel(resourceLabelVo);
-					}
+			}
+			//如果有标签id列表
+			if (labelIds!=null&&labelIds.size()!=0){
+				Integer resourceid = record.getResourceid();
+				if (resourceid==null||resourceNewsService.selectByPrimaryKey(resourceid)==null){
+					return JsonResult.getFail("未获取到正确的resourceid");
+				}
+				ResourceLabelVo resourceLabelVo = new ResourceLabelVo();
+				resourceLabelVo.setCreator(admin.getAdminId());
+				resourceLabelVo.setUpdater(admin.getAdminId());
+				resourceLabelVo.setCreatetime(currentTime);
+				resourceLabelVo.setUpdatetime(currentTime);
+				resourceLabelVo.setTableid(3);
+				resourceLabelVo.setResourceid(resourceid);
+				resourceLabelVo.setStatus(1);
+				resourceLabelService.deleteResourceLabelByResourceId(resourceid);
+				for (Integer labelId : labelIds) {
+					resourceLabelVo.setLabelid(labelId);
+					resourceLabelService.addResourceLabel(resourceLabelVo);
 				}
 			}
 			resourceNewsService.updateResourceNews(record);
